requirements.txt
python-pptx>=0.6.23
python-dotenv>=1.0.1
requests>=2.31.0
pywin32>=306; platform_system == "Windows"
